Move assay actions to assay/actions.ts from base/actions.ts
